requirements.txt
keystoneauth1==2.20.0
requests-futures==0.9.7
py-zabbix=1.1.3
xlsxwriter==1.0.0